Sets up initial gem to patch shoulda-matchers
Since we're taking advantage of storing UUIDs in MySQL as `binary`, the shoulda-matchers has an issue when the UUID ends in an `f`.  This has lead to some flakey tests.  [We made an effort to fix this upstream](https://github.com/thoughtbot/shoulda-matchers/pull/1159), which doesn't appear to be gaining traction.  This extracts the fix into a gem so we don't have to support a hard fork for this.

* Adds rubocop and fix linter issues
* Adds circleci config and ingore gem build artifacts
* Adds CHANGELOG.md with latest change
* Ensure database.yml is correctly setup for CI
* Ensure CI builds with mysql to ensure tests run
* Adds license to gemspec and updates readme with license info
* Prepare README for public consumption
* Update gem publishing to use Rubygems
* Removes dox-style in favor of raw rubocop configuration to make open-source compatable