Sets up initial gem to patch shoulda-matchers
Since we're taking advantage of storing UUIDs in MySQL as `binary`, the shoulda-matchers has an issue when the UUID ends in an `f`.  This has lead to some flakey tests.  [We made an effort to fix this upstream](https://github.com/thoughtbot/shoulda-matchers/pull/1159), which doesn't appear to be gaining traction.  This extracts the fix into a gem so we don't have to support a hard fork for this.

Pivotal Story: https://www.pivotaltracker.com/story/show/164603806
Merges #1
LGTM given by: @bsimpson